effort on systems that didnt pan out
unfortunately this session was a bit of a waste of time, I tried to implement a system that would have agents back up from eachother if they got too close that did not really work out, then I tried to implement an airlaunch that didnt pan out. the big contributions I did today were implementing a system to track who each agent is targeting and being targeted by to allow for that agent to choose to dodge or block, adding some notes for future plans, and resizing some hitboxes.